rename to TableKit
Thanks a lot to my greatest friend, Alexander Nikishin
