+++ TableKit/TableKit.h
@@ -1,0 +1,19 @@
+//
+//  TableKit.h
+//  TableKit
+//
+//  Created by Max Sokolov on 11/06/16.
+//  Copyright © 2016 Max Sokolov. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for TableKit.
+FOUNDATION_EXPORT double TableKitVersionNumber;
+
+//! Project version string for TableKit.
+FOUNDATION_EXPORT const unsigned char TableKitVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <TableKit/PublicHeader.h>
+
+
