+++ TableKitDemo/Resources/Assets/Assets.xcassets/AppIcon.appiconset/Contents.json
