+++ TableKitDemo/Resources/Info.plist
