--- TabletDemo/Classes/Presentation/Views/StoryboardImageTableViewCell.swift
@@ -1,33 +1,0 @@
-//
-//  StoryboardImageTableViewCell.swift
-//  TabletDemo
-//
-//  Created by Max Sokolov on 24/05/16.
-//  Copyright © 2016 Tablet. All rights reserved.
-//
-
-import UIKit
-import Tablet
-
-class StoryboardImageTableViewCell: UITableViewCell, ConfigurableCell {
-    
-    typealias T = String
-    
-    @IBOutlet var titleLabel: UILabel!
-    @IBOutlet var subtitleLabel: UILabel!
-    @IBOutlet var customImageView: UIImageView!
-    
-    func configure(string: T) {
-        
-        titleLabel.text = string
-        subtitleLabel.text = "Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.Copyright © 2016 Tablet. All rights reserved.1"
-    }
-
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        
-        contentView.layoutIfNeeded()
-        
-        subtitleLabel.preferredMaxLayoutWidth = subtitleLabel.bounds.size.width
-    }
-}
